+++ software/distro/setup/base-os/cleanup.sh
@@ -1,6 +1,11 @@
 #!/bin/bash -eux
 # Cleanup removes unnecessary files from the operating system for a smaller and more secure disk image.
 
+# Remove unnecessary packages:
+sudo apt-get remove -y gcc
+sudo apt-get remove -y gcc-10 || true
+sudo apt-get remove -y gcc-12 || true
+
 # Clean up any unnecessary apt files:
 sudo apt-get autoremove -y
 sudo apt-get clean -y