+++ software/distro/setup/base-os/forklift/setup-booted.sh
@@ -1,11 +1,4 @@
 #!/bin/bash -eux
-# The Forklift pallet github.com/PlanktoScope/pallet-standard provides the standard configuration of
-# Forklift package deployments of Docker containerized applications, OS config files, and systemd
-# system services for the PlanktoScope software distribution. This script integrates that pallet
-# into the PlanktoScope OS's filesystem by installing Forklift and providing some systemd units
-# which set up bind mounts and overlay filesystems to bootstrap the configs managed by Forklift.
-
-config_files_root=$(dirname $(realpath $BASH_SOURCE))
 
 # Prepare to apply the local pallet
 
@@ -15,7 +8,7 @@ config_files_root=$(dirname $(realpath $BASH_SOURCE))
 # script here (even though it works after the script finishes, before rebooting):
 FORKLIFT="forklift"
 if [ -S /var/run/docker.sock ] && \
-  ! sudo -E docker ps 2>&1 > /dev/null && \
+  ! sudo -E docker info && \
   ! sudo systemctl start docker.socket docker.service
 then
   echo "Error: couldn't start docker!"
@@ -31,19 +24,9 @@ then
   sudo iptables -L || sudo lsmod
   exit 1
 fi
-if ! docker ps 2>&1 > /dev/null; then
+if ! docker info 2>&1 > /dev/null; then
   FORKLIFT="sudo -E forklift"
 fi
-
-# Move container images used by the local pallet from the pre-cache to the Docker daemon
-
-echo "Downloading temporary tools to load pre-downloaded container images into Docker..."
-tmp_bin="$(mktemp -d --tmpdir=/tmp bin.XXXXXXX)"
-"$config_files_root/download-rush.sh" "$tmp_bin"
-
-echo "Loading pre-downloaded container images into Docker..."
-forklift plt ls-img | "$tmp_bin/rush" \
-  "$config_files_root/transfer-precached-image.sh" {} "$HOME/.cache/forklift/containers/docker-archives"
 
 # Applying the staged pallet (i.e. making Docker instantiate all the containers) significantly
 # decreases first-boot time, by up to 30 sec for github.com/PlanktoScope/pallet-standard.