+++ software/distro/setup/base-os/tools/install.sh
@@ -7,7 +7,7 @@
 # of package versions for existing packages if we run the script at different times. Also, it causes
 # some weirdness with the Docker installation.
 sudo apt-get install -y -o DPkg::Lock::Timeout=60 -o Dpkg::Progress-Fancy=0 \
-  vim byobu git curl parallel
+  vim byobu git curl
 
 # Install some tools for dealing with captive portals
 sudo apt-get install -y -o Dpkg::Progress-Fancy=0 \